Move examples into their own directory.
Having `gorename(1)` just work is a "good thing(tm)."
